+++ src/common/components/index.js
@@ -1,2 +1,3 @@
 export * from './CreateItemField';
+export * from './FieldClaimingDate';
 export * from './LineLocationsView';