+++ src/common/utils/index.js
@@ -1,3 +1,4 @@
+export * from './getClaimingIntervalFromDate';
 export * from './getDehydratedPiece';
 export * from './getItemById';
 export * from './getReceivingPieceItemStatus';