+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * View piece status change log. Refs UIREC-305.
 * Add CSV export options for new piece statuses. Refs UIREC-306.
 * Add the "Acquisition units" protected field to the receiving title form. Refs UIREC-295.
+* Delay claim action for piece record. Refs UIREC-303.
+* Send claim action for piece record. Refs UIREC-304.
 
 ## [4.0.0](https://github.com/folio-org/ui-receiving/tree/v4.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v3.0.0...v4.0.0)